Upgrade from djs v11 to djs v12.

+++ commands/Moderation/ban.js
@@ -22,13 +22,17 @@ module.exports = {
             const success = functions.newEmbed()
                 .setAuthor('Ban')
                 .setDescription('User has successfully been banned!')
-                .addField('User', member.user.tag, false)
-                .addField('Moderator', message.author.tag, false)
-                .addField('Reason', args[1] ? args.slice(1).join(' ') : '-', false)
-                .setThumbnail(member.user.displayAvatarURL);
+                .addFields(
+                    [
+                        { name: 'User', value: member.user.tag },
+                        { name: 'Moderator', value: message.author.tag },
+                        { name: 'Reason', value: args[1] ? args.slice(1).join(' ') : '-' }
+                    ]
+                )
+                .setThumbnail(member.user.displayAvatarURL({ size: 256, dynamic: true }));
 
             message.channel.send(success);
-            message.client.channels.get(config.modLogChannel).send(success);
+            message.client.channels.cache.get(config.modLogChannel).send(success);
             member.send(success.setDescription(`You have been banned from ${message.guild.name}.`))
                 .catch(() => {
                     return;
@@ -44,12 +48,15 @@ module.exports = {
             return ban(member);
 
         const menu = functions.newEmbed();
-        const memberObj = message.guild.members.filter(m => m.user.username.toLowerCase().includes(args[0].toLowerCase()));
+        const memberObj = message.guild.members.cache.filter(m => m.user.username.toLowerCase().includes(args[0].toLowerCase()));
         const members = [];
-        memberObj.forEach(m => members.push([m.user.username, m.user.displayAvatarURL, m]));
+        memberObj.forEach(m => members.push([m.user.username, m.user.displayAvatarURL({ size: 256, dynamic: true }), m]));
 
         let i = 0;
-        menu.setTitle(`Selected user: ${members[i][0]}`).setThumbnail(members[i][1]).setFooter(`${i + 1}/${members.length}`).setDescription('<a:loading:679334829359366165> *In the process of banning a user...*\n\nUse ⬅️/➡️ to navigate pages.\nUse ✅ to confirm and kick the selected user.\nUse ❌ to cancel any time.');
+        menu
+            .setTitle(`Selected user: ${members[i][0]}`)
+            .setThumbnail(members[i][1]).setFooter(`${i + 1}/${members.length}`)
+            .setDescription('<a:loading:679334829359366165> *In the process of banning a user...*\n\nUse ⬅️/➡️ to navigate pages.\nUse ✅ to confirm and kick the selected user.\nUse ❌ to cancel any time.');
 
         const msg = await message.channel.send(menu);
         await msg.react('⬅️');
@@ -61,9 +68,10 @@ module.exports = {
         const spam = msg.createReactionCollector((reaction, user) => user.id !== message.author.id, { time: 1000 * 60 * 5 });
 
         spam.on('collect', r => {
-            r.users.filter(user => user.id !== message.client.user.id).forEach(user => r.remove(user));
+            r.users.cache.filter(user => user.id !== message.client.user.id).forEach(user => r.users.remove(user));
         });
         collector.on('collect', r => {
+            r.users.cache.filter(user => user.id !== message.client.user.id).forEach(user => r.users.remove(user));
             switch (r.emoji.name) {
                 case '⬅️':
                     if (i === 0)
@@ -90,10 +98,10 @@ module.exports = {
                 default:
                     break;
             }
-            r.users.filter(user => user.id !== message.client.user.id).forEach(user => r.remove(user));
         });
         collector.on('end', () => {
-            msg.delete();
+            if (msg)
+                msg.delete();
             if (!member || member === 'reactions')
                 return;
             ban(member);
